add function to collect data from form
